Move Composer into a package
